requirements.txt
# Guarda este contenido en un archivo llamado requirements.txt

python-telegram-bot
requests